added search hp test
