+++ src/test/java/Tests/brain/acoustics/CheckAcousticSystemsCategory.java
@@ -1,0 +1,41 @@
+package Tests.brain.acoustics;
+
+import PageObject.BasePage;
+import PageObject.brain.com.ua.BrainAcousticsPage;
+import PageObject.brain.com.ua.BrainHomePage;
+import PageObject.brain.com.ua.BrainOrderPage;
+import PageObject.brain.com.ua.BrainProductPage;
+import Tests.TestInit;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class CheckAcousticSystemsCategory extends TestInit {
+
+    @Test
+    public void checkAcousticSystemsCategory() {
+
+        goToSite("https://brain.com.ua/ukr/");
+        BasePage page = new BasePage(driver);
+        page.implicitlyWait();
+
+        BrainHomePage homePage = new BrainHomePage(driver);
+        homePage.getAcoustics().click();
+
+        BrainAcousticsPage acousticsPage = new BrainAcousticsPage(driver);
+        acousticsPage.getAcousticSystems().click();
+        homePage.getFirstModel().click();
+
+        BrainProductPage productPage = new BrainProductPage(driver);
+        sleep(5);
+        productPage.getBuyBtn().click();
+        productPage.getCheckOut().click();
+
+        BrainOrderPage orderPage = new BrainOrderPage(driver);
+        orderPage.privat24().click();
+        orderPage.newPost().click();
+        orderPage.recipientData();
+        orderPage.cancelPurchase();
+
+        Assert.assertTrue(orderPage.cartIsEmpty().isDisplayed());
+    }
+}
