+++ src/test/java/Tests/brain/acoustics/CheckMicrophonesCategory.java
@@ -1,15 +1,18 @@
-package Tests.brain;
+package Tests.brain.acoustics;
 
 import PageObject.BasePage;
-import PageObject.brain.com.ua.*;
+import PageObject.brain.com.ua.BrainAcousticsPage;
+import PageObject.brain.com.ua.BrainHomePage;
+import PageObject.brain.com.ua.BrainOrderPage;
+import PageObject.brain.com.ua.BrainProductPage;
 import Tests.TestInit;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class CheckHeadphonesCategory extends TestInit {
+public class CheckMicrophonesCategory extends TestInit {
 
     @Test
-    public void checkHeadphonesCategory() {
+    public void checkMicrophonesCategory() {
 
         goToSite("https://brain.com.ua/ukr/");
         BasePage page = new BasePage(driver);
@@ -19,7 +22,7 @@ public class CheckHeadphonesCategory extends TestInit {
         homePage.getAcoustics().click();
 
         BrainAcousticsPage acousticsPage = new BrainAcousticsPage(driver);
-        acousticsPage.getHeadphones().click();
+        acousticsPage.getMicrophones().click();
         homePage.getFirstModel().click();
 
         BrainProductPage productPage = new BrainProductPage(driver);
